+++ apps/website/CHANGELOG.md
@@ -1,5 +1,19 @@
 # website
 
+## 2.3.3
+
+### Patch Changes
+
+- Updated dependencies [31d05ed5]
+- Updated dependencies [dddb3d1a]
+- Updated dependencies [68995558]
+- Updated dependencies [ebba8404]
+- Updated dependencies [91412d77]
+- Updated dependencies [a3180b02]
+- Updated dependencies [91412d77]
+  - @saas-ui/react@2.4.0
+  - @saas-ui/forms@2.3.0
+
 ## 2.3.2
 
 ### Patch Changes