+++ packages/saas-ui-hotkeys/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @saas-ui/hotkeys
 
+## 2.1.6
+
+### Patch Changes
+
+- Updated dependencies [31d05ed5]
+- Updated dependencies [dddb3d1a]
+- Updated dependencies [ebba8404]
+- Updated dependencies [91412d77]
+- Updated dependencies [a3180b02]
+- Updated dependencies [91412d77]
+  - @saas-ui/core@2.2.0
+
 ## 2.1.5
 
 ### Patch Changes