+++ packages/saas-ui-modals/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @saas-ui/modals
 
+## 2.1.7
+
+### Patch Changes
+
+- Updated dependencies [68995558]
+- Updated dependencies [a3180b02]
+  - @saas-ui/forms@2.3.0
+
 ## 2.1.6
 
 ### Patch Changes